net: update False Start blacklist for 32-bit indexes.
(Needed for the next list update.)

BUG=none
TEST=net_unittests

http://codereview.chromium.org/4599002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65376 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a70a02d6af0441f7eed5a1c8996a7e505a87aced

+++ net/base/ssl_false_start_blacklist.h
@@ -81,7 +81,7 @@ class SSLFalseStartBlacklist {
 
   // kHashTable contains an offset into |kHashData| for each bucket. The
   // additional element at the end contains the length of |kHashData|.
-  static const uint16 kHashTable[kBuckets + 1];
+  static const uint32 kHashTable[kBuckets + 1];
   // kHashData contains the contents of the hash table. |kHashTable| indexes
   // into this array. Each bucket consists of zero or more, 8-bit length
   // prefixed strings. Each string is a DNS name in dotted form. For a given
